Created common file for test helpers and switched to using `utils.promisify(fs.readFile)`.
